refactor: change verificationController To UserVerificationSpecification
