+++ src/main/java/lems/cowshed/api/controller/user/UserVerificationController.java
@@ -1,4 +1,4 @@
-package lems.cowshed.api.controller.mail;
+package lems.cowshed.api.controller.user;
 
 import lems.cowshed.api.controller.CommonResponse;
 import lems.cowshed.api.controller.dto.verification.request.MailVerificationRequest;
@@ -12,30 +12,30 @@ import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 
-import static lems.cowshed.domain.mail.code.CodeType.*;
+import static lems.cowshed.domain.mail.code.CodeType.SIGN_UP;
 
 @RequiredArgsConstructor
-@RequestMapping("/verification")
+@RequestMapping("/user/verification")
 @RestController
-public class VerificationApiController implements VerificationSpecification {
+public class UserVerificationController implements UserVerificationSpecification {
 
     private final UserService userService;
     private final MailService mailService;
     private final CodeFinder codeFinder;
 
-    @PostMapping("/mail/{mail}")
-    public CommonResponse<VerificationResultInfo> verifyMailAndSendCode(@PathVariable String mail){
-        if(userService.isExistEmail(mail)){
+    @PostMapping("/email/{email}")
+    public CommonResponse<VerificationResultInfo> verifyEmailAndSendCode(@PathVariable String email){
+        if(userService.isExistEmail(email)){
             return CommonResponse.success(VerificationResultInfo.of(false));
         }
 
         String code = codeFinder.findCodeFrom(SIGN_UP);
-        mailService.sendCodeToMail(SIGN_UP, Mail.of(mail, code));
+        mailService.sendCodeToMail(SIGN_UP, Mail.of(email, code));
         return CommonResponse.success(VerificationResultInfo.of(true));
     }
 
-    @GetMapping("/mail")
-    public CommonResponse<VerificationResultInfo> verifyMail(@RequestBody MailVerificationRequest request){
+    @GetMapping("/email/code")
+    public CommonResponse<VerificationResultInfo> verifyEmail(@RequestBody MailVerificationRequest request){
         return CommonResponse.success(mailService.verifyMail(request.toMail(), LocalDateTime.now()));
     }
